+++ ast/ast.go
@@ -2,8 +2,6 @@ package ast
 
 import (
 	"bytes"
-
-	"github.com/cszczepaniak/monkey/token"
 )
 
 type Node interface {
@@ -11,19 +9,11 @@ type Node interface {
 	String() string
 }
 
-type Statement interface {
-	Node
-	statementNode()
-}
-
-type Expression interface {
-	Node
-	expressionNode()
-}
-
 type Program struct {
 	Statements []Statement
 }
+
+var _ Node = (*Program)(nil)
 
 func (p *Program) TokenLiteral() string {
 	if len(p.Statements) > 0 {
@@ -31,245 +21,10 @@ func (p *Program) TokenLiteral() string {
 	}
 	return ""
 }
-
 func (p *Program) String() string {
 	var out bytes.Buffer
 	for _, s := range p.Statements {
 		out.WriteString(s.String())
 	}
-	return out.String()
-}
-
-type LetStatement struct {
-	Token token.Token
-	Name  *Identifier
-	Value Expression
-}
-
-func (ls *LetStatement) statementNode() {}
-func (ls *LetStatement) TokenLiteral() string {
-	return ls.Token.Literal
-}
-
-func (ls *LetStatement) String() string {
-	var out bytes.Buffer
-
-	out.WriteString(ls.TokenLiteral() + " ")
-	out.WriteString(ls.Name.String())
-	out.WriteString(" = ")
-	if ls.Value != nil {
-		out.WriteString(ls.Value.String())
-	}
-	out.WriteString(";")
-
-	return out.String()
-}
-
-type Identifier struct {
-	Token token.Token
-	Value string
-}
-
-func (i *Identifier) expressionNode() {}
-func (i *Identifier) TokenLiteral() string {
-	return i.Token.Literal
-}
-
-func (i *Identifier) String() string {
-	return i.Value
-}
-
-type ReturnStatement struct {
-	Token       token.Token
-	ReturnValue Expression
-}
-
-func (rs *ReturnStatement) statementNode() {}
-func (rs *ReturnStatement) TokenLiteral() string {
-	return rs.Token.Literal
-}
-
-func (rs *ReturnStatement) String() string {
-	var out bytes.Buffer
-
-	out.WriteString(rs.TokenLiteral() + " ")
-	if rs.ReturnValue != nil {
-		out.WriteString(rs.ReturnValue.String())
-	}
-	out.WriteString(";")
-
-	return out.String()
-}
-
-type ExpressionStatement struct {
-	Token      token.Token
-	Expression Expression
-}
-
-func (es *ExpressionStatement) statementNode() {}
-func (es *ExpressionStatement) TokenLiteral() string {
-	return es.Token.Literal
-}
-
-func (es *ExpressionStatement) String() string {
-	if es.Expression != nil {
-		return es.Expression.String()
-	}
-	return ""
-}
-
-type IntegerLiteral struct {
-	Token token.Token
-	Value int64
-}
-
-func (il *IntegerLiteral) expressionNode() {}
-func (il *IntegerLiteral) TokenLiteral() string {
-	return il.Token.Literal
-}
-
-func (il *IntegerLiteral) String() string {
-	return il.Token.Literal
-}
-
-type PrefixExpression struct {
-	Token    token.Token
-	Operator string
-	Right    Expression
-}
-
-func (pe *PrefixExpression) expressionNode() {}
-func (pe *PrefixExpression) TokenLiteral() string {
-	return pe.Token.Literal
-}
-func (pe *PrefixExpression) String() string {
-	return `(` + pe.Operator + pe.Right.String() + `)`
-}
-
-type InfixExpression struct {
-	Token    token.Token
-	Operator string
-	Right    Expression
-	Left     Expression
-}
-
-func (ie *InfixExpression) expressionNode() {}
-func (ie *InfixExpression) TokenLiteral() string {
-	return ie.Token.Literal
-}
-func (ie *InfixExpression) String() string {
-	return `(` + ie.Left.String() + ` ` + ie.Operator + ` ` + ie.Right.String() + `)`
-}
-
-type BooleanLiteral struct {
-	Token token.Token
-	Value bool
-}
-
-func (bl *BooleanLiteral) expressionNode() {}
-func (bl *BooleanLiteral) TokenLiteral() string {
-	return bl.Token.Literal
-}
-func (bl *BooleanLiteral) String() string {
-	return bl.Token.Literal
-}
-
-type IfExpression struct {
-	Token       token.Token
-	Condition   Expression
-	Consequence *BlockStatement
-	Alternative *BlockStatement
-}
-
-func (ie *IfExpression) expressionNode() {}
-func (ie *IfExpression) TokenLiteral() string {
-	return ie.Token.Literal
-}
-func (ie *IfExpression) String() string {
-	var out bytes.Buffer
-
-	out.WriteString(`if`)
-	out.WriteString(ie.Condition.String())
-	out.WriteString(` `)
-	out.WriteString(ie.Consequence.String())
-
-	if ie.Alternative != nil {
-		out.WriteString(`else `)
-		out.WriteString(ie.Alternative.String())
-	}
-
-	return out.String()
-}
-
-type BlockStatement struct {
-	Token      token.Token
-	Statements []Statement
-}
-
-func (bs *BlockStatement) statementNode() {}
-func (bs *BlockStatement) TokenLiteral() string {
-	return bs.Token.Literal
-}
-func (bs *BlockStatement) String() string {
-	var out bytes.Buffer
-
-	out.WriteString(`{ `)
-	for _, s := range bs.Statements {
-		out.WriteString(s.String())
-		out.WriteString(`; `)
-	}
-	out.WriteString(`}`)
-
-	return out.String()
-}
-
-type FunctionLiteral struct {
-	Token token.Token
-	Args  []*Identifier
-	Body  *BlockStatement
-}
-
-func (fl *FunctionLiteral) expressionNode() {}
-func (fl *FunctionLiteral) TokenLiteral() string {
-	return fl.Token.Literal
-}
-func (fl *FunctionLiteral) String() string {
-	var out bytes.Buffer
-	out.WriteString(`fn(`)
-	for i, arg := range fl.Args {
-		out.WriteString(arg.String())
-		if i < len(fl.Args)-1 {
-			out.WriteString(`, `)
-		}
-	}
-	out.WriteString(`) `)
-	out.WriteString(fl.Body.String())
-
-	return out.String()
-}
-
-type CallExpression struct {
-	Token    token.Token
-	Function Expression
-	Args     []Expression
-}
-
-func (ce *CallExpression) expressionNode() {}
-func (ce *CallExpression) TokenLiteral() string {
-	return ce.Token.Literal
-}
-func (ce *CallExpression) String() string {
-	var out bytes.Buffer
-
-	out.WriteString(ce.Function.String())
-	out.WriteString(`(`)
-	for i, arg := range ce.Args {
-		out.WriteString(arg.String())
-		if i < len(ce.Args)-1 {
-			out.WriteString(`, `)
-		}
-	}
-	out.WriteString(`)`)
-
 	return out.String()
 }
